Login and navigation operations moved to helper classes
