+++ test_add_group.py
@@ -2,6 +2,8 @@
 from selenium.webdriver.firefox.webdriver import WebDriver
 import unittest
 from group import Group
+from navigation import Navigation
+from login import Auth
 
 
 def is_alert_present(wd):
@@ -19,36 +21,21 @@ class TestAddGroup(unittest.TestCase):
     
     def test_add_group(self):
         wd = self.wd
-        self.open_home_page(wd)
-        self.login(wd, username="admin", password="secret")
-        self.open_groups_page(wd)
+        Navigation().open_home_page(wd)
+        Auth().login(wd, username="admin", password="secret")
+        Navigation().open_groups_page(wd)
         self.create_group(wd, Group(name="group_002", header="another_header", footer="another_footer"))
-        self.return_to_groups_page(wd)
-        self.logout(wd)
+        Navigation().open_groups_page(wd)
+        Auth().logout(wd)
 
     def test_add_empty_group(self):
         wd = self.wd
-        self.open_home_page(wd)
-        self.login(wd, username="admin", password="secret")
-        self.open_groups_page(wd)
+        Navigation().open_home_page(wd)
+        Auth().login(wd, username="admin", password="secret")
+        Navigation().open_groups_page(wd)
         self.create_group(wd, Group(name="", header="", footer=""))
-        self.return_to_groups_page(wd)
-        self.logout(wd)
-
-    def open_home_page(self, wd):
-        wd.get("http://localhost/addressbook/")
-
-    def login(self, wd, username, password):
-        wd.find_element_by_name("user").click()
-        wd.find_element_by_name("user").clear()
-        wd.find_element_by_name("user").send_keys(username)
-        wd.find_element_by_name("pass").click()
-        wd.find_element_by_name("pass").clear()
-        wd.find_element_by_name("pass").send_keys(password)
-        wd.find_element_by_xpath("//form[@id='LoginForm']/input[3]").click()
-
-    def open_groups_page(self, wd):
-        wd.find_element_by_link_text("groups").click()
+        Navigation().open_groups_page(wd)
+        Auth().logout(wd)
 
     def create_group(self, wd, group):
         # open group creation form
@@ -65,12 +52,6 @@ class TestAddGroup(unittest.TestCase):
         wd.find_element_by_name("group_footer").send_keys(group.footer)
         # submit group creation
         wd.find_element_by_name("submit").click()
-
-    def return_to_groups_page(self, wd):
-        wd.find_element_by_link_text("group page").click()
-
-    def logout(self, wd):
-        wd.find_element_by_link_text("Logout").click()
 
     def tearDown(self):
         self.wd.quit()
